Sistemati problemi CORS! (2)

+++ src/main/java/com/example/springbootrentcar/controller/PrenotazioneController.java
@@ -13,6 +13,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDate;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
 
 @RestController
@@ -47,8 +49,13 @@ public class PrenotazioneController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deletePrenotazione(@PathVariable int id) {
-        prenotazioneService.deletePrenotazione(id);
+    public void deletePrenotazione(@PathVariable int id) throws Exception {
+        LocalDate dataInizio = LocalDate.parse(prenotazioneService.getPrenotazione(id).getDataInizio());
+        if (dataInizio.until(LocalDate.now(), ChronoUnit.DAYS) > 2) {
+            prenotazioneService.deletePrenotazione(id);
+        } else {
+            throw new Exception("Errore, non è possibile cancellare entro due giorni dalla prenotazione");
+        }
     }
 
     private Utente getUtenteBySession() {
